These lines iterate over the indices of the correlation matrix. The outer loop iterates over the columns of the correlation matrix using the corr_matrix.columns attribute. The inner loop iterates over the indices up to the current column index.
Inside the loop, the code checks if the absolute value of the correlation between the columns at index (i, j) is greater than the threshold value. If it is, the column name corresponding to the index i is stored in the variable colname. This column name is then added to the col_corr set.